backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
openai>=1.0.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
